Remove unnecessary grouping parentheses.
These are occasionally helpful, so not all of them were removed, and the
jshint option that warns about them was disabled. Basic arithmetic
precedence (PEMDAS) should be clear without the use of parentheses. Also
they are not needed around return values. However, they can be helpful and
clarifying in other cases where the precedence isn't straightforward.